Prevent setting value, if mapping has more than one entry
Multiple entries in a mapping mean, that if the first entry was not found,
the next entry should be used to determine the value of a field.
For that reason we cannot set the value of the field just yet and we need to keep
through the data until a value was found or the value remains empty.