Do not add skipped-parent-item in child model
Child models do not have a last-update-field and have added a skipped-parent-item every time
they were processed. This has led to the fact that child relations were not deleted correctly.
If a parent is skipped because it has not changed, the parent itself will add its own skipped id
to the child model and don't even process any relations in these child models.